--- lib/nonnative/definition.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module Nonnative
-  class Definition
-    attr_accessor :process
-    attr_accessor :timeout
-    attr_accessor :port
-    attr_accessor :file
-  end
-end
